Moved tests that fail on Windows to a separate class
